api/requirements.txt
# PhantomProxy API - Enterprise Killer Stack
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
sqlmodel==0.0.22
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# Redis + Celery
redis==5.2.0
celery[redis]==5.4.0

# Auth (Keycloak/Zitadel compatible)
python-jose[cryptography]==3.3.0
httpx==0.27.2

# Observability (optional)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
